src/third_party/blink/web_tests/external/wpt/css/selectors/attribute-selectors/attribute-case/cssom-expected.txt

This is a testharness.js-based test.
Found 72 tests; 40 PASS, 32 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS [foo="bar"] /* sanity check */ insertRule
PASS [foo="bar"] /* sanity check */ getting CSSRule#cssText
PASS [foo="bar"] /* sanity check */ getting CSSStyleRule#selectorText
PASS [foo="bar"] /* sanity check */ setting CSSStyleRule#selectorText
PASS [foo="bar"] /* sanity check */ insertRule in @media
PASS [foo="bar"] /* sanity check */ getting CSSRule#cssText in @media
PASS [foo="bar"] /* sanity check */ getting CSSStyleRule#selectorText in @media
PASS [foo="bar"] /* sanity check */ setting CSSStyleRule#selectorText in @media
PASS [foo="bar" i] insertRule
PASS [foo="bar" i] getting CSSRule#cssText
PASS [foo="bar" i] getting CSSStyleRule#selectorText
PASS [foo="bar" i] setting CSSStyleRule#selectorText
PASS [foo="bar" i] insertRule in @media
PASS [foo="bar" i] getting CSSRule#cssText in @media
PASS [foo="bar" i] getting CSSStyleRule#selectorText in @media
PASS [foo="bar" i] setting CSSStyleRule#selectorText in @media
PASS [foo="bar" /**/ i] insertRule
PASS [foo="bar" /**/ i] getting CSSRule#cssText
PASS [foo="bar" /**/ i] getting CSSStyleRule#selectorText
PASS [foo="bar" /**/ i] setting CSSStyleRule#selectorText
PASS [foo="bar" /**/ i] insertRule in @media
PASS [foo="bar" /**/ i] getting CSSRule#cssText in @media
PASS [foo="bar" /**/ i] getting CSSStyleRule#selectorText in @media
PASS [foo="bar" /**/ i] setting CSSStyleRule#selectorText in @media
PASS [foo="bar"/**/i] insertRule
PASS [foo="bar"/**/i] getting CSSRule#cssText
PASS [foo="bar"/**/i] getting CSSStyleRule#selectorText
PASS [foo="bar"/**/i] setting CSSStyleRule#selectorText
PASS [foo="bar"/**/i] insertRule in @media
PASS [foo="bar"/**/i] getting CSSRule#cssText in @media
PASS [foo="bar"/**/i] getting CSSStyleRule#selectorText in @media
PASS [foo="bar"/**/i] setting CSSStyleRule#selectorText in @media
PASS [*|foo="bar" i] insertRule
PASS [*|foo="bar" i] getting CSSRule#cssText
PASS [*|foo="bar" i] getting CSSStyleRule#selectorText
PASS [*|foo="bar" i] setting CSSStyleRule#selectorText
PASS [*|foo="bar" i] insertRule in @media
PASS [*|foo="bar" i] getting CSSRule#cssText in @media
PASS [*|foo="bar" i] getting CSSStyleRule#selectorText in @media
PASS [*|foo="bar" i] setting CSSStyleRule#selectorText in @media
FAIL [foo="bar" s] insertRule Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[foo="bar" s] {}'.
FAIL [foo="bar" s] getting CSSRule#cssText Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[foo="bar" s] {}'.
FAIL [foo="bar" s] getting CSSStyleRule#selectorText Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[foo="bar" s] {}'.
FAIL [foo="bar" s] setting CSSStyleRule#selectorText assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [foo="bar" s] insertRule in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[foo="bar" s] {}' is invalid and cannot be parsed.
FAIL [foo="bar" s] getting CSSRule#cssText in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[foo="bar" s] {}' is invalid and cannot be parsed.
FAIL [foo="bar" s] getting CSSStyleRule#selectorText in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[foo="bar" s] {}' is invalid and cannot be parsed.
FAIL [foo="bar" s] setting CSSStyleRule#selectorText in @media assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [foo="bar" /**/ s] insertRule Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[foo="bar" /**/ s] {}'.
FAIL [foo="bar" /**/ s] getting CSSRule#cssText Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[foo="bar" /**/ s] {}'.
FAIL [foo="bar" /**/ s] getting CSSStyleRule#selectorText Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[foo="bar" /**/ s] {}'.
FAIL [foo="bar" /**/ s] setting CSSStyleRule#selectorText assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [foo="bar" /**/ s] insertRule in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[foo="bar" /**/ s] {}' is invalid and cannot be parsed.
FAIL [foo="bar" /**/ s] getting CSSRule#cssText in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[foo="bar" /**/ s] {}' is invalid and cannot be parsed.
FAIL [foo="bar" /**/ s] getting CSSStyleRule#selectorText in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[foo="bar" /**/ s] {}' is invalid and cannot be parsed.
FAIL [foo="bar" /**/ s] setting CSSStyleRule#selectorText in @media assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [foo="bar"/**/s] insertRule Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[foo="bar"/**/s] {}'.
FAIL [foo="bar"/**/s] getting CSSRule#cssText Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[foo="bar"/**/s] {}'.
FAIL [foo="bar"/**/s] getting CSSStyleRule#selectorText Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[foo="bar"/**/s] {}'.
FAIL [foo="bar"/**/s] setting CSSStyleRule#selectorText assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [foo="bar"/**/s] insertRule in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[foo="bar"/**/s] {}' is invalid and cannot be parsed.
FAIL [foo="bar"/**/s] getting CSSRule#cssText in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[foo="bar"/**/s] {}' is invalid and cannot be parsed.
FAIL [foo="bar"/**/s] getting CSSStyleRule#selectorText in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[foo="bar"/**/s] {}' is invalid and cannot be parsed.
FAIL [foo="bar"/**/s] setting CSSStyleRule#selectorText in @media assert_equals: expected "[foo=\"bar\" s]" but got "foobar"
FAIL [*|foo="bar" s] insertRule Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[*|foo="bar" s] {}'.
FAIL [*|foo="bar" s] getting CSSRule#cssText Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[*|foo="bar" s] {}'.
FAIL [*|foo="bar" s] getting CSSStyleRule#selectorText Failed to execute 'insertRule' on 'CSSStyleSheet': Failed to parse the rule '[*|foo="bar" s] {}'.
FAIL [*|foo="bar" s] setting CSSStyleRule#selectorText assert_equals: expected "[*|foo=\"bar\" s]" but got "foobar"
FAIL [*|foo="bar" s] insertRule in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[*|foo="bar" s] {}' is invalid and cannot be parsed.
FAIL [*|foo="bar" s] getting CSSRule#cssText in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[*|foo="bar" s] {}' is invalid and cannot be parsed.
FAIL [*|foo="bar" s] getting CSSStyleRule#selectorText in @media Failed to execute 'insertRule' on 'CSSGroupingRule': the rule '[*|foo="bar" s] {}' is invalid and cannot be parsed.
FAIL [*|foo="bar" s] setting CSSStyleRule#selectorText in @media assert_equals: expected "[*|foo=\"bar\" s]" but got "foobar"
Harness: the test ran to completion.
